allow re-run if process data file is not there

+++ modules/assim.batch/R/pda.emulator.R
@@ -56,9 +56,11 @@ pda.emulator <- function(settings, external.data = NULL, external.priors = NULL,
   # load inputs with neff if this is another round
   if(!run.normal){
     external_data_path <- file.path(settings$outdir, paste0("external.", settings$assim.batch$ensemble.id, ".Rdata"))
-    load(external_data_path)
-    # and delete the file afterwards because it will be re-written with a new ensemble id in the end
-    file.remove(external_data_path)
+    if(exists(external_data_path)){
+      load(external_data_path)
+      # and delete the file afterwards because it will be re-written with a new ensemble id in the end
+      file.remove(external_data_path)
+    } 
   }
   
   ## will be used to check if multiplicative Gaussian is requested
